:rocket: Moving detail view to dynamic route
Solving auto-login bug with metamask due to weird nextjs's deal with dynamic routes
