+++ package.json
@@ -20,7 +20,7 @@
 		"firebase": "9.8.4",
 		"framer-motion": "^5",
 		"ipfs-http-client": "^57.0.1",
-		"next": "12.0.8",
+		"next": "12.2.0",
 		"qrcode.react": "^1.0.1",
 		"react": "17.0.2",
 		"react-dom": "17.0.2",
